# SubTracker-main/docs/api-notes/subtrack-openapi.yaml
openapi: 3.0.3
info:
  title: SubTrack API
  version: 0.1.0
  description: MVP backend contract for SubTrack demo.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/v1/auth/register:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation or business error
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ApiErrorResponse'
                  - $ref: '#/components/schemas/ValidationErrorResponse'
  /api/v1/auth/login:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/v1/analytics/summary:
    get:
      tags: [analytics]
      parameters:
        - in: query
          name: period
          required: true
          schema:
            type: string
            enum: [month, year]
      responses:
        '200':
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummaryResponse'
        '400':
          description: Invalid period
  /api/v1/analytics/by-category:
    get:
      tags: [analytics]
      parameters:
        - in: query
          name: period
          required: true
          schema:
            type: string
            enum: [month, year]
      responses:
        '200':
          description: Analytics by category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsByCategoryResponse'
        '400':
          description: Invalid period
...
  /api/v1/analytics/usage:
    get:
      tags: [analytics]
      parameters:
        - in: query
          name: period
          required: true
          schema:
            type: string
            enum: [month, year]
        - in: query
          name: subscriptionId
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Usage analytics by active subscriptions for period
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsUsageResponse'
        '400':
          description: Invalid period or subscriptionId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/forecast:
    get:
      tags: [forecast]
      deprecated: true
      description: Deprecated compatibility alias. Uses the same calculation as /api/v1/analytics/forecast.
      parameters:
        - in: query
          name: months
          required: false
          deprecated: true
          schema:
            type: integer
            default: 12
          description: Deprecated. Kept for backward compatibility and ignored by calculation logic.
      responses:
        '200':
          description: Forecast (legacy shape)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastResponse'
  /api/v1/consents/imports/{provider}:
    get:
      tags: [consents]
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [GMAIL]
      responses:
        '200':
          description: Current consent state for import provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportConsentStatusResponse'
        '400':
          description: Invalid or unsupported provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/consents/imports/{provider}/grant:
    post:
      tags: [consents]
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [GMAIL]
      responses:
        '200':
          description: Consent granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportConsentStatusResponse'
        '400':
          description: Invalid or unsupported provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/consents/imports/{provider}/revoke:
    post:
      tags: [consents]
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [GMAIL]
      responses:
        '200':
          description: Consent revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportConsentStatusResponse'
        '400':
          description: Invalid or unsupported provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/imports/start:
    post:
      tags: [imports]
      description: Requires active consent for requested provider. Grant via /api/v1/consents/imports/{provider}/grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportStartRequest'
            examples:
              success:
                summary: Fully successful import
                value:
                  provider: GMAIL
                  messages:
                    - externalId: gmail-msg-success-001
                      from: billing@netflix.com
                      subject: Netflix renewal notice
                      body: "Your Netflix renewal is scheduled for 2026-03-20. Amount: 9.99 USD"
                      receivedAt: 2026-03-08T10:00:00Z
                    - externalId: gmail-msg-success-002
                      from: billing@netflix.com
                      subject: Netflix renewal notice
                      body: "Your Netflix renewal is scheduled for 2026-04-20. Amount: 9.99 USD"
                      receivedAt: 2026-03-08T10:05:00Z
              partial:
                summary: Partial import with parser errors
                value:
                  provider: GMAIL
                  messages:
                    - externalId: gmail-msg-partial-001
                      from: billing@netflix.com
                      subject: Netflix renewal notice
                      body: "Your Netflix renewal is scheduled for 2026-03-22. Amount: 9.99 USD"
                      receivedAt: 2026-03-08T11:00:00Z
                    - externalId: gmail-msg-partial-002
                      from: billing@netflix.com
                      subject: Netflix renewal notice
                      body: "Amount: 9.99 USD"
                      receivedAt: 2026-03-08T11:01:00Z
                    - externalId: gmail-msg-partial-003
                      from: no-reply@example.com
                      subject: Monthly statement
                      body: Payment details enclosed
                      receivedAt: 2026-03-08T11:02:00Z
              fail:
                summary: Full failure (all messages invalid for parser)
                value:
                  provider: GMAIL
                  messages:
                    - externalId: gmail-msg-fail-001
                      from: no-reply@example.com
                      subject: Your receipt
                      body: "Amount: 9.99 USD"
                      receivedAt: 2026-03-08T12:00:00Z
                    - externalId: gmail-msg-fail-002
                      from: billing@netflix.com
                      subject: Netflix notice
                      body: Renewal amount 9.99 USD
                      receivedAt: 2026-03-08T12:01:00Z
      responses:
        '200':
          description: Import finished (sync MVP)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResultResponse'
        '400':
          description: Invalid provider or broken request payload
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ApiErrorResponse'
                  - $ref: '#/components/schemas/ValidationErrorResponse'
              examples:
                unsupportedProvider:
                  summary: Provider is not supported in MVP
                  value:
                    message: only GMAIL provider is supported in MVP
                validationError:
                  summary: DTO validation error
                  value:
                    message: Validation failed
                    errors:
                      messages[0].externalId: externalId may contain only letters, digits, dot, underscore, colon and hyphen
                      messages[0].from: from must be a valid email address
        '403':
          description: Import consent is missing for provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportConsentErrorResponse'
              examples:
                consentRequired:
                  summary: Import blocked without active consent
                  value:
                    code: IMPORT_CONSENT_REQUIRED
                    message: active consent is required before import for provider GMAIL
                    provider: GMAIL
  /api/v1/imports/{provider}/sync:
    post:
      tags: [imports]
      description: Synchronizes supported mailbox provider and runs import pipeline.
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
            enum: [GMAIL]
      responses:
        '200':
          description: Mailbox sync finished (sync MVP)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResultResponse'
        '400':
          description: Invalid provider or mailbox flow is disabled by feature flag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Import consent is missing for provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportConsentErrorResponse'
        '409':
          description: Mailbox is not connected or requires OAuth reconnect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailboxFlowErrorResponse'
...
  /api/v1/imports/{id}:
    get:
      tags: [imports]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Import job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResultResponse'
        '400':
          description: Import job not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          maxLength: 100
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
          nullable: true
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserResponse'
...
    AnalyticsSummaryResponse:
      type: object
      properties:
        period:
          type: string
          enum: [month, year]
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        totalAmount:
          type: number
          format: double
        activeSubscriptions:
          type: integer
          format: int64
        currency:
          type: string
          nullable: true
    AnalyticsByCategoryResponse:
      type: object
      properties:
        period:
          type: string
          enum: [month, year]
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        totalAmount:
          type: number
          format: double
        currency:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsCategoryItemResponse'
    AnalyticsCategoryItemResponse:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
          format: double
        sharePercent:
          type: number
          format: double
        subscriptionsCount:
          type: integer
          format: int32
...
    AnalyticsUsageResponse:
      type: object
      properties:
        period:
          type: string
          enum: [month, year]
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        totalSignals:
          type: integer
          format: int64
        activeSubscriptions:
          type: integer
          format: int64
        subscriptionsWithSignals:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsUsageItemResponse'
    AnalyticsUsageItemResponse:
      type: object
      properties:
        subscriptionId:
          type: integer
          format: int64
        serviceName:
          type: string
        category:
          type: string
        signalsCount:
          type: integer
          format: int64
        lastSignalAt:
          type: string
          format: date-time
          nullable: true
    ForecastResponse:
      type: object
      properties:
        monthlyForecast:
          type: number
          format: double
        yearlyForecast:
          type: number
          format: double
    ImportConsentStatusResponse:
      type: object
      properties:
        provider:
          type: string
          enum: [GMAIL]
        status:
          type: string
          enum: [GRANTED, REVOKED, NOT_GRANTED]
        scope:
          type: string
          example: MAIL_IMPORT_READ
        grantedAt:
          type: string
          format: date-time
          nullable: true
        revokedAt:
          type: string
          format: date-time
          nullable: true
        integrationStatus:
          type: string
          example: ACTIVE
    ImportConsentErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: IMPORT_CONSENT_REQUIRED
        message:
          type: string
        provider:
          type: string
          enum: [GMAIL]
    MailboxFlowErrorResponse:
      type: object
      properties:
        code:
          type: string
          enum: [MAILBOX_CONNECTION_REQUIRED, MAILBOX_REAUTH_REQUIRED]
        message:
          type: string
        provider:
          type: string
          enum: [GMAIL]
    ImportStartRequest:
      type: object
      required: [provider, messages]
      properties:
        provider:
          type: string
          enum: [GMAIL]
          description: Only GMAIL is supported in MVP import flow and requires active consent.
        messages:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/MailMessageRequest'
    MailMessageRequest:
      type: object
      required: [externalId, from, subject, body, receivedAt]
      properties:
        externalId:
          type: string
          maxLength: 255
          pattern: '^[A-Za-z0-9._:-]{1,255}$'
        from:
          type: string
          format: email
          maxLength: 255
        subject:
          type: string
          maxLength: 500
        body:
          type: string
          maxLength: 10000
        receivedAt:
          type: string
          format: date-time
    ImportResultResponse:
      type: object
      properties:
        jobId:
          type: integer
          format: int64
        provider:
          type: string
        status:
          type: string
          enum: [IN_PROGRESS, COMPLETED, COMPLETED_WITH_ERRORS, FAILED]
        processed:
          type: integer
          format: int32
        created:
          type: integer
          format: int32
        skipped:
          type: integer
          format: int32
        errors:
          type: integer
          format: int32
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        errorItems:
          type: array
          items:
            $ref: '#/components/schemas/ImportErrorItemResponse'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ImportItemResultResponse'
      example:
        jobId: 42
        provider: GMAIL
        status: COMPLETED_WITH_ERRORS
        processed: 3
        created: 1
        skipped: 1
        errors: 1
        startedAt: 2026-03-08T10:00:00Z
        finishedAt: 2026-03-08T10:00:01Z
        errorItems:
          - externalId: gmail-msg-003
            reason: parser could not extract next billing date (expected YYYY-MM-DD)
        items:
          - externalId: gmail-msg-001
            status: IMPORTED
            reason: null
            sourceProvider: netflix.com
            serviceName: Netflix
            amount: 9.99
            currency: USD
            billingPeriod: MONTHLY
            nextBillingDate: 2026-03-20
            category: Entertainment
            receivedAt: 2026-03-08T10:00:00Z
          - externalId: gmail-msg-003
            status: PARSE_ERROR
            reason: parser could not extract next billing date (expected YYYY-MM-DD)
            sourceProvider: netflix.com
            serviceName: Netflix
            amount: null
            currency: null
            billingPeriod: null
            nextBillingDate: null
            category: Entertainment
            receivedAt: 2026-03-08T10:01:00Z
    ImportItemResultResponse:
      type: object
      properties:
        externalId:
          type: string
        status:
          type: string
          enum: [IMPORTED, SKIPPED, PARSE_ERROR, UNSUPPORTED]
        reason:
          type: string
          nullable: true
        sourceProvider:
          type: string
          nullable: true
        serviceName:
          type: string
          nullable: true
        amount:
          type: number
          format: double
          nullable: true
        currency:
          type: string
          nullable: true
        billingPeriod:
          type: string
          enum: [WEEKLY, MONTHLY, QUARTERLY, YEARLY]
          nullable: true
        nextBillingDate:
          type: string
          format: date
          nullable: true
        category:
          type: string
          nullable: true
        receivedAt:
          type: string
          format: date-time
          nullable: true
    ImportErrorItemResponse:
      type: object
      properties:
        externalId:
          type: string
        reason:
          type: string
...